+++ docs/user_guide/changelog.rst
@@ -24,6 +24,7 @@ Bug Fixes
 ---------
 
 - Fix thread safety issue in LogContext.
+- Fix torch.clamp with integer tensors
 
 v2.4 (Poplar SDK 2.4)
 =====================